+++ auth-db/queries/users/create.ts
@@ -1,10 +1,10 @@
 import { DbKey, queryWrapper } from "@saflib/drizzle-sqlite3";
-import { SelectUser } from "../../types.ts";
-import { NewUser } from "../../types.ts";
+import type { SelectUser } from "../../types.ts";
+import type { NewUser } from "../../types.ts";
 import { authDbManager } from "../../instances.ts";
 import { EmailConflictError } from "../../errors.ts";
 import { users } from "../../schema.ts";
-import { ReturnsError } from "@saflib/monorepo";
+import type { ReturnsError } from "@saflib/monorepo";
 
 export const create = queryWrapper(
   async (
