+++ auth-db/queries/users/index.ts
@@ -1,1 +1,2 @@
 export * from "./create.ts";
+export * from "./get-all.ts";
